added weather display to Map View and refactored
