+++ frontend/src/app/results/_components/_MapViewComponents/MapView.tsx
@@ -1,23 +1,24 @@
 import React, { useState, useEffect, useRef } from "react";
 import axios from "axios";
-import {
-  GoogleMap,
-  LoadScript,
-  Marker,
-  useLoadScript,
-} from "@react-google-maps/api";
 import MapModal from "./MapModal";
+import { FlightsData } from "../../_interfaces/flightsInterfaces";
+import MapContainer from "./MapContainer";
 
-const MapComponent = ({ flightsData }: { flightsData: any }) => {
-  const [latitude, setLatitude] = useState(null); //marker requests
+interface MapComponentProps {
+  flightsData: FlightsData
+}
+
+const MapComponent = ({ flightsData }: MapComponentProps) => {
+  // marker requests
+  const [latitude, setLatitude] = useState(null);
   const [longitude, setLongitude] = useState(null);
   const [isMapModalOpen, setIsMapModalOpen] = useState(true);
-  const [MarkerCenter, setCenter] = useState<{ lat: any; lng: any }>({
+  const [markerCenter, setCenter] = useState<{ lat: any; lng: any }>({
     lat: 40.4948,
     lng: -3.5793544,
   });
   const [markers, setMarkers] = useState<{ lat: any; lng: any }[]>([]);
-  const [MapCenter, setMapCenter] = useState<{ lat: any; lng: any }>({
+  const [mapCenter, setMapCenter] = useState<{ lat: any; lng: any }>({
     lat: 40.4948,
     lng: -3.5793544,
   });
@@ -113,48 +114,19 @@ const MapComponent = ({ flightsData }: { flightsData: any }) => {
               isOpen={isMapModalOpen}
               onClose={closeMapModal}
               flights={data.flights}
+              weather={data.destinationWeather}
               updateMarkerCenter={updateMarkercenter}
               onAirportCodesReceived={handleAirportCodes}
             />
           </div>
           <div className="w-3/4 justify-center">
-            <div>
-              <LoadScript
-                googleMapsApiKey={`${process.env.NEXT_PUBLIC_GOOGLE_MAPS_API_KEY}`}
-              >
-                <GoogleMap
-                  mapContainerStyle={mapContainerStyle}
-                  center={MarkerCenter}
-                  zoom={zoom}
-                >
-                  <Marker position={MarkerCenter} />
-                  {markers.map((marker, index) => (
-                    <Marker
-                      key={index}
-                      position={{
-                        lat: marker.lat,
-                        lng: marker.lng,
-                      }}
-                    />
-                  ))}
-                </GoogleMap>
-              </LoadScript>
-            </div>
+            <MapContainer
+              mapContainerStyle={mapContainerStyle}
+              markerCenter={markerCenter}
+              zoom={zoom}
+              markers={markers}
+            />
           </div>
-        </div>
-        <div className="flex flex-row justify-center items-center rounded-lg bg-gray-200 h-[50px] border border-gray-300 gap-x-20">
-          <button
-            className="bg-gradient-to-r from-blue-500 to-green-500 text-white px-4 py-2 rounded-lg font-bold w-[300px]"
-            onClick={openMapModal}
-          >
-            Flight results
-          </button>
-          <button className="bg-gradient-to-r from-blue-500 to-green-500 text-white px-4 py-2 rounded-lg font-bold w-[300px]">
-            Things to do
-          </button>
-          <button className="bg-gradient-to-r from-blue-500 to-green-500 text-white px-4 py-2 rounded-lg font-bold w-[300px]">
-            TV Reccomendations
-          </button>
         </div>
       </div>
     </>
