+++ frontend/src/app/results/_components/_MapViewComponents/ModalStyles.css
@@ -1,5 +1,5 @@
 .modal-scrollable-content {
-  max-height: 575px;
+  max-height: 430px;
   overflow-y: auto;
 }
